requirements.txt
python-dotenv
openai 
langchain
langgraph
pandas
tavily-python
langchain-openai